+++ src/raspberrypi/rasctl/rasctl_display.cpp
@@ -9,7 +9,7 @@
 
 #include "rascsi_interface.pb.h"
 #include "rasutil.h"
-#include "display.h"
+#include "rasctl_display.h"
 #include <iostream>
 #include <list>
